+++ finstack/valuations/tests/instruments/trs/test_fi_index_trs.rs
@@ -11,6 +11,7 @@ use finstack_core::money::Money;
 use finstack_valuations::cashflow::traits::CashflowProvider;
 use finstack_valuations::instruments::common::traits::Instrument;
 use finstack_valuations::instruments::TrsSide;
+use rust_decimal::Decimal;
 
 // ================================================================================================
 // Construction and Validation Tests
@@ -42,7 +43,7 @@ fn test_fi_index_trs_builder_custom_params() {
     // Assert
     assert_eq!(trs.id.as_str(), "CUSTOM-FI-TRS-001");
     assert_eq!(trs.notional.amount(), 25_000_000.0);
-    assert_eq!(trs.financing.spread_bp, 150.0);
+    assert_eq!(trs.financing.spread_bp, Decimal::from(150));
     assert_eq!(trs.side, TrsSide::PayTotalReturn);
 }
